pyproject.toml
[tool.poetry]
name = "happybudget"
version = "1.0.0"
description = "Backend Application for HappyBudget Application"
authors = ["nickmflorin <nickmflorin@gmail.com>"]
repository = "https://github.com/nickmflorin/happybudget-api"
homepage = "https://github.com/nickmflorin/happybudget-api"
include = [
  "README.md",
  "pyproject.toml"
]

[tool.poetry.dependencies]
python = "^3.8"
dj-database-url = "0.5.0"
Django = "4.0"
tox = "^3.20.1"
python-dotenv = "^0.15.0"
pytest = "^6.2.2"
django-cors-headers = "^3.7.0"
cryptography = "^3.4.6"
django-ratelimit = "^3.0.1"
django-model-utils = "^4.1.1"
django-polymorphic = "^3.0.0"
psycopg2 = "^2.8.6"
django-timezone-field = "^4.1.1"
django-extensions = "^3.1.1"
psycopg2-binary = "^2.8.6"
gunicorn = "^20.0.4"
python-dateutil = "^2.8.1"
boto3 = "^1.17.32"
django-storages = "^1.11.1"
Pillow = "^8.1.2"
django-colorful = "^1.3"
xhtml2pdf = "^0.2.5"
rest-framework-generic-relations = "^2.0.0"
django-grappelli = "^2.15.1"
django-libsass = "^0.8"
watchtower = "^1.0.6"
python_http_client = "^3.3.3"
raven = "^6.10.0"
django-redis = "^5.0.0"
redis-py-cluster = "^2.1.3"
djangorestframework = "^3.12.4"
djangorestframework_simplejwt = "^5.0.0"
stripe = "^2.61.0"
sib-api-v3-sdk = "^7.3.0"
django-nested-admin = "^3.4.0"
tqdm = "^4.62.3"
plaid-python = "^9.1.0"
fontawesomefree = "^6.1.1"

[tool.poetry.extras]
docs = [
  "sphinx",
  "sphinx-rtd-theme",
  "sphinxcontrib.documentedlist",
  "sphinxcontrib.httpdomain",
  "sphinxcontrib.httpexample"
]
debug = [ "ipython", "ipdb" ]

[tool.poetry.dev-dependencies]
pytest = "^6.1.1"
coverage = "^4.5"
flake8 = "^3.7"
pytest-freezegun = "^0.4.2"
factory-boy = "3.0.1"
pytest-django = "3.10.0"
responses = "^0.12.1"
mock = "4.0.2"
pylint = "2.6.0"
pytest-cov = "^2.10.1"
pytest-pythonpath = "^0.7.3"
flake8_formatter_junit_xml = "^0.0.6"
ipdb = "0.13.4"
autopep8 = "^1.5.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
